Refactor trajectory export logic and clean up baseline tests
